implement user.getByClerkId handler

+++ api/migrations/000006_create-questions-downvotes-table.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE questionsDownvotes (
-  questionId INT REFERENCES questions(_id) ON DELETE CASCADE NOT NULL,
-  userId INT REFERENCES users(_id) ON DELETE CASCADE NOT NULL,
+  questionId BIGINT REFERENCES questions(_id) ON DELETE CASCADE NOT NULL,
+  userId BIGINT REFERENCES users(_id) ON DELETE CASCADE NOT NULL,
   UNIQUE (questionId, userId)
 );
